use autoloader-dev in composer for TestBase files
