+++ composer.json
@@ -287,7 +287,11 @@
   },
   "autoload-dev": {
     "psr-4": {
-      "Drupal\\Tests\\hs_bugherd\\": "docroot/modules/humsci/hs_bugherd/tests"
+      "Drupal\\Tests\\hs_basic_page\\": "docroot/modules/humsci/hs_basic_page/tests",
+      "Drupal\\Tests\\hs_bugherd\\": "docroot/modules/humsci/hs_bugherd/tests",
+      "Drupal\\Tests\\hs_courses_importer\\": "docroot/modules/humsci/hs_courses/modules/hs_courses_importer/tests",
+      "Drupal\\Tests\\hs_field_helpers\\": "docroot/modules/humsci/hs_field_helpers/tests",
+      "Drupal\\Tests\\hs_table_filter\\": "docroot/modules/humsci/hs_table_filter/tests"
     }
   },
   "scripts": {
